create common module for sylo types
