+++ pallet/common/src/lib.rs
@@ -674,17 +674,5 @@ impl Migrator for () {
 		Ok(())
 	}
 }
-pub trait SyloDataVerificationProvider {
-	type AccountId;
-	type StringLimit: Get<u32>;
 
-	// fn get_validation_record(
-	// 	author: &Self::AccountId,
-	// 	data_id: &BoundedVec<u8, Self::StringLimit>,
-	// ) -> bool
-
-	fn validation_record_exists(
-		author: &Self::AccountId,
-		data_id: &BoundedVec<u8, Self::StringLimit>,
-	) -> bool;
-}
+pub mod sylo;
